@@ -3,7 +3,7 @@ description: Demonstrates how to use the flutter_scan_bluetooth plugin.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
@@ -12,12 +12,12 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.6
+
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
   flutter_scan_bluetooth:
     path: ../
 
