@@ -18,5 +18,5 @@ flutter:
         pluginClass: FlutterScanBluetoothPlugin
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=2.0.0"
+  sdk: ">=2.15.1 <4.0.0"
+  flutter: ">=2.5.0"
