@@ -70,7 +70,7 @@ class _MyAppState extends State<MyApp> {
             Padding(
               padding: const EdgeInsets.all(8.0),
               child: Center(
-                child: RaisedButton(child: Text('Check permissions'), onPressed: () async {
+                child: ElevatedButton(child: Text('Check permissions'), onPressed: () async {
                   try {
                       await _bluetooth.requestPermissions();
                       print('All good with perms');
